Mock-based tests for block query response API
